chore: test cases replaced
